fix minor details
