+++ src/components/home/LikeToDo/LikeToDoItem.tsx
@@ -5,20 +5,14 @@ type LikeToDoItemProps = {
   item: LikeTodo
   iconActive: string
   accordionName: string
-  setAccordionName: Dispatch<SetStateAction<string>>
 }
 export default function LikeToDoItem({
   item,
   iconActive,
   accordionName,
-  setAccordionName,
 }: LikeToDoItemProps) {
   return (
-    <div
-      key={item.title}
-      className='border-none text-left'
-      onMouseOver={() => setAccordionName(item.title)}
-      onMouseLeave={() => setAccordionName('')}>
+    <div key={item.title} className='border-none text-left'>
       <h3
         className={`text-2xl font-bold transition ${
           item.title === iconActive || item.title === accordionName
@@ -28,11 +22,9 @@ export default function LikeToDoItem({
         {item.title}
       </h3>
       {(item.title === accordionName || item.title === iconActive) && (
-        <div className='transform transition-transform'>
-          <p className='text-justify text-white opacity-40 hover:opacity-80'>
-            {item.value}
-          </p>
-        </div>
+        <p className='transform text-justify text-base text-white opacity-40 transition-transform hover:opacity-80'>
+          {item.value}
+        </p>
       )}
     </div>
   )
